requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
openai>=1.0.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
tiktoken>=0.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector>=0.2.0
litellm>=1.0.0
python-dotenv>=1.0.0

# Optional dependencies for better performance
torch>=2.0.0
transformers>=4.30.0 